user access control

+++ app/Http/Controllers/PostsController.php
@@ -6,6 +6,16 @@ use Illuminate\Http\Request;
 use App\Post;
 class PostsController extends Controller
 {
+        /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' =>['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -69,6 +79,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post=Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !==$post->user->id){
+            return redirect('/posts')->with('error','Un-authorized Page');
+        }
+
         return View('posts.edit')->with('post',$post);
     }
 
@@ -102,6 +118,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post=Post::find($id);
+
+         // Check for correct user
+         if(auth()->user()->id !==$post->user->id){
+            return redirect('/posts')->with('error','Un-authorized Page');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success','Post deleted');
     }
